requirements.txt
openai>=1.54.0
requests>=2.32.0
python-dotenv>=1.0.1
gpiozero>=2.0; platform_system == "Linux"
